% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{SpatCovarEst}
\alias{SpatCovarEst}
\title{Estimates covariance between point pattern and spatial covariance with
a kernel estimation procedure. Bandwidth selection is employed to selecet
the optimal bandwidth.}
\usage{
SpatCovarEst(
  X,
  Z,
  R,
  r,
  b_init = NULL,
  b_limit = NULL,
  grid = NULL,
  fast = TRUE
)
}
\arguments{
\item{X}{A point pattern represented by a ppp object.}

\item{Z}{The spatial covariate represented as a marked ppp object where the
coordinates represents the spatial locations where the coviaraite is
measured, and the marks represents the measured values.}

\item{R}{The range over which the error of the esimator is assessed. This
is used to calculate the Mean Integrated Squared Error (MISE) used for
bandwidth selection. Must be a positive number.}

\item{r}{Vector of distance(s) for which to esimate the spatial covariance.}

\item{b_init}{Initial value for optimisation of the bandwidth. If left empty,
R/2 is used. If Grid is non-NULL, b_init should be left as NULL.}

\item{b_limit}{Search interval for b. If NULL, MISE is minimized over b in [0,R].}

\item{grid}{Grid to evaulate the MISE over. The bandwidth with the lowest MISE
will then be selected for the mixed moment estimator. If set to NULL the MISE
is found with numeric optimsation optimised. Using grid can sometimes save
considerable computational resources, but unless this is relevant consideration,
it is recommended to leave grid = NULL.}

\item{fast}{Should the leave-one-out cross-validation use a fast approximation? 
True by default.}
}
\value{
Returns a list containing the estimated covariances (c0), the
distances at which the covariances are estimated (r), and the selected
bandwidth (b).
}
\description{
Estimates covariance between point pattern and spatial covariance with
a kernel estimation procedure. Bandwidth selection is employed to selecet
the optimal bandwidth.
}
