% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{SpatCovarEstFixed}
\alias{SpatCovarEstFixed}
\title{Estimates covariance between point pattern and spatial covariance with
a kernel estimation procedure for a user-specified bandwidth.}
\usage{
SpatCovarEstFixed(X, Z, r, b)
}
\arguments{
\item{X}{A point pattern represented by a ppp object.}

\item{Z}{The spatial covariate represented as a marked ppp object where the
coordinates represents the spatial locations where the coviaraite is
measured, and the marks represents the measured values.}

\item{r}{Vector of distance(s) for which to esimate the spatial covariance.}

\item{b}{The selected bandwidth.}
}
\value{
Returns a list containing the estimated covariances (c0) and the
distances at which the covariances are estimated (r).
}
\description{
Estimates covariance between point pattern and spatial covariance with
a kernel estimation procedure for a user-specified bandwidth.
}
