% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{SpatCovarEst}
\alias{SpatCovarEst}
\title{Estimates covariance between point pattern and spatial covariance with
a kernel estimation procedure. Bandwidth selection is employed to selecet
the optimal bandwidth.}
\usage{
SpatCovarEst(X, Z, R, r)
}
\arguments{
\item{X}{A point pattern represented by a ppp object.}

\item{Z}{The spatial covariate represented as a marked ppp object where the
coordinates represents the spatial locations where the coviaraite is
measured, and the marks represents the measured values.}

\item{R}{The range over which the error of the esimator is assessed. This
is used to calculate the Mean Integrated Squared Error (MISE) used for
bandwidth selection. Must be a positive number.}

\item{r}{Vector of distance(s) for which to esimate the spatial covariance.}
}
\value{
Returns a list containing the estimated covariances (c0), the
distances at which the covariances are estimated (r), and the selected
bandwidth (b).
}
\description{
Estimates covariance between point pattern and spatial covariance with
a kernel estimation procedure. Bandwidth selection is employed to selecet
the optimal bandwidth.
}
